[Add] - %x and %p now both working with precision and field_w

+++ libft/ft_printf/output/print_identifier.c
@@ -29,7 +29,7 @@ void					print_identifier(t_pf_item *pfi, va_list args)
 	else if (pfi->cspecs->o)
 		print_oct(pfi, va_arg(args, long int));
 	else if (pfi->cspecs->x || pfi->cspecs->lg_x)
-		print_hex(pfi, va_arg(args, long int));
+		print_hex(pfi, va_arg(args, int));
 	else if (pfi->cspecs->p)
 		print_address(pfi, va_arg(args, long int));
 }
